sintab: Proper casts for integer conversion
We need to cast through a signed fixed-point type before casting to unsigned, else negative values may be truncated (observed with rustc 1.64.0 nightly).

Also includes an automatic update to Cargo.lock.

+++ sintab/src/main.rs
@@ -6,11 +6,10 @@ use std::io::Write;
 fn main() {
     let output_file_name = env::args().skip(1).next().unwrap();
 
-    let bytes =
-        (0..256)
+    let bytes = (0..256)
         .map(|i| {
             let f = (i as f64) / 128.0 * PI;
-            (f.sin() * 64.0/* + (f * 3.0).sin() * 32.0*/) as u8
+            (f.sin() * 64.0/* + (f * 3.0).sin() * 32.0*/) as i8 as u8
         })
         .collect::<Vec<_>>();
